+++ src/components/SplitButton/index.ts
@@ -1,0 +1,12 @@
+export {
+  SplitButton,
+  SplitButtonAction,
+  SplitButtonTrigger,
+  SplitButtonContent,
+} from './SplitButton';
+export type {
+  SplitButtonProps,
+  SplitButtonActionProps,
+  SplitButtonTriggerProps,
+  SplitButtonContentProps,
+} from './SplitButton';